Dedupe code: DBus message interpretation
Finally moved the part that interprets a DBus return/event message completely away from the code of both conditions and events that rely on such interpretation. This should end the code deduplication effort, because for Lua the amount of duplicate code is minimal.
