+++ src/condition/dbus_cond.rs
@@ -16,7 +16,6 @@ use async_std::task;
 use zbus;
 use zbus::zvariant;
 
-use std::convert::TryFrom;
 use std::str::FromStr;
 use serde_json::value::Value;
 
@@ -1193,33 +1192,6 @@ impl Condition for DbusMethodCondition {
         //       the zbus::blocking option might be considered for further
         //       developement as well as rebuilding this service as real async
 
-        // a helper to apply a given operator to two values without clutter;
-        // for simplicity sake the `Match` operator will just evaluate to
-        // `false` here, instead of generating an error: the `Err()` case would
-        // clutter the code for numerical comparisons uslessly, as we also know
-        // that the test are built only via `load_cfgmap`, and that it only
-        // admits 'match' for regular expressions; the `Contains` operator also
-        // evaluates to `false` here since this function only compares args
-        // that are `PartialOrd+PartialEq`, and arrays are not
-        fn _oper<T: PartialOrd+PartialEq>(op: &ParamCheckOperator, o1: T, o2: T) -> bool {
-            match op {
-                ParamCheckOperator::Equal => o1 == o2,
-                ParamCheckOperator::NotEqual => o1 != o2,
-                ParamCheckOperator::Less => o1 < o2,
-                ParamCheckOperator::LessEqual => o1 <= o2,
-                ParamCheckOperator::Greater => o1 > o2,
-                ParamCheckOperator::GreaterEqual => o1 >= o2,
-                ParamCheckOperator::Match => false,
-                ParamCheckOperator::Contains => false,
-                ParamCheckOperator::NotContains => false,
-            }
-        }
-
-        // the following function allows for better readability
-        fn _contained_in<T: Containable>(v: &T, a: &zvariant::Value) -> bool {
-            v.is_contained_in(a)
-        }
-
         // panic here if the bus name is incorrect: should have been fixed
         // when the condition was configured and constructed
         async fn _get_connection(bus: &str) -> zbus::Result<zbus::Connection> {
@@ -1342,656 +1314,31 @@ impl Condition for DbusMethodCondition {
         }
 
         // now check method result in the same way as in signal message
-        // (see `_start_service` in `dbus_event` for details)
-        let mut verified = self.param_checks_all;
+        let verified;
         if let Some(checks) = &self.param_checks {
-            if let Ok(mbody) = message.unwrap().clone().body::<zvariant::Structure>() {
-                // the label is set to make sure that we can break out from
-                // any nested loop on shortcut evaluation condition (that is
-                // when all condition had to be true and at least one is false
-                // or when one true condition is sufficient and we find it)
-                // or when an error occurs, which implies evaluation to false
-                'params: for ck in checks.iter() {
-                    let argnum = ck.index.get(0);
-                    if let Some(argnum) = argnum {
-                        match argnum {
-                            ParameterIndex::Integer(x) => {
-                                if *x >= mbody.fields().len() as u64 {
-                                    self.log(
-                                        LogType::Warn,
-                                        LOG_WHEN_PROC,
-                                        LOG_STATUS_FAIL,
-                                        &format!("could not retrieve result: index {x} out of range"),
-                                    );
-                                    verified = false;
-                                    break 'params;
-                                }
-                                let mut field_value = mbody.fields().get(*x as usize).unwrap();
-                                for x in 1 .. ck.index.len() {
-                                    match ck.index.get(x).unwrap() {
-                                        ParameterIndex::Integer(i) => {
-                                            let i = *i as usize;
-                                            match field_value {
-                                                zvariant::Value::Array(f) => {
-                                                    if i >= f.len() {
-                                                        self.log(
-                                                            LogType::Warn,
-                                                            LOG_WHEN_PROC,
-                                                            LOG_STATUS_FAIL,
-                                                            &format!("could not retrieve result: index {i} out of range"),
-                                                        );
-                                                    }
-                                                    // if something is wrong here, either the test
-                                                    // or the next "parameter shift" will go berserk
-                                                    field_value = &f[i];
-                                                }
-                                                zvariant::Value::Structure(f) => {
-                                                    let f = f.fields();
-                                                    if i >= f.len() {
-                                                        self.log(
-                                                            LogType::Warn,
-                                                            LOG_WHEN_PROC,
-                                                            LOG_STATUS_FAIL,
-                                                            &format!("could not retrieve result: index {i} out of range"),
-                                                        );
-                                                    }
-                                                    if let Some(v) = f.get(i) {
-                                                        field_value = &v;
-                                                    } else {
-                                                        self.log(
-                                                            LogType::Warn,
-                                                            LOG_WHEN_PROC,
-                                                            LOG_STATUS_FAIL,
-                                                            &format!("could not retrieve result: index {i} provided no value"),
-                                                        );
-                                                    }
-                                                }
-                                                _ => {
-                                                    self.log(
-                                                        LogType::Warn,
-                                                        LOG_WHEN_PROC,
-                                                        LOG_STATUS_FAIL,
-                                                        &format!("could not retrieve result using index {i}"),
-                                                    );
-                                                    verified = false;
-                                                    break 'params;
-                                                }
-                                            }
-                                        }
-                                        // even though there would be the possibility to explicitly indicate an ObjectPath
-                                        // via an '\o' prefix (as we do for values), since object paths really are strings
-                                        // that adhere to a certain format, the conversion is automatically done for the
-                                        // cases where an ObjectPath is required in place of a string - just logging that
-                                        // a malformed string was configured as index where a well-formed object path had
-                                        // to be provided; the following code directly matching the signature beginning is
-                                        // in fact ugly as hell, however a nicer implementation might come with more recent
-                                        // releases of zbus, which provide enum variants for signature nested structures
-                                        ParameterIndex::String(s) => {
-                                            let s = s.as_str();
-                                            match field_value {
-                                                zvariant::Value::Dict(f) => {
-                                                    // in order to match either strings or object paths, match key signature
-                                                    let m= match key_signature(f).as_str() {
-                                                        "s" => {
-                                                            let k = zvariant::Str::from(s);
-                                                            f.get(&k)
-                                                        }
-                                                        "o" => {
-                                                            let res = zvariant::ObjectPath::try_from(s);
-                                                            if res.is_err() {
-                                                                self.log(
-                                                                    LogType::Warn,
-                                                                    LOG_WHEN_PROC,
-                                                                    LOG_STATUS_FAIL,
-                                                                    &format!("could not retrieve result: index `{s}` should be an object path"),
-                                                                );
-                                                                verified = false;
-                                                                break 'params;
-                                                            } else {
-                                                                let k = res.unwrap().to_owned();
-                                                                f.get(&k)
-                                                            }
-                                                        }
-                                                        _ =>{
-                                                            self.log(
-                                                                LogType::Warn,
-                                                                LOG_WHEN_PROC,
-                                                                LOG_STATUS_FAIL,
-                                                                &format!("could not retrieve result: index `{s}` not matching dictionary key type"),
-                                                            );
-                                                            verified = false;
-                                                            break 'params;
-                                                        }
-                                                    };
-                                                    field_value = match m {
-                                                        Ok(fv) => {
-                                                            if let Some(fv) = fv {
-                                                                fv
-                                                            } else {
-                                                                self.log(
-                                                                    LogType::Warn,
-                                                                    LOG_WHEN_PROC,
-                                                                    LOG_STATUS_FAIL,
-                                                                    &format!("could not retrieve result: index `{s}` invalid"),
-                                                                );
-                                                                verified = false;
-                                                                break 'params;
-                                                            }
-                                                        },
-                                                        Err(_) => {
-                                                            self.log(
-                                                                LogType::Warn,
-                                                                LOG_WHEN_PROC,
-                                                                LOG_STATUS_FAIL,
-                                                                &format!("could not retrieve result using index `{s}`"),
-                                                            );
-                                                            verified = false;
-                                                            break 'params;
-                                                        }
-                                                    }
-                                                }
-                                                _ => {
-                                                    self.log(
-                                                        LogType::Warn,
-                                                        LOG_WHEN_PROC,
-                                                        LOG_STATUS_FAIL,
-                                                        &format!("could not retrieve parameter using index `{s}`"),
-                                                    );
-                                                    verified = false;
-                                                    break 'params;
-                                                }
-                                            }
-                                        }
-                                    }
-                                }
+            self.log(
+                LogType::Debug,
+                LOG_WHEN_PROC,
+                LOG_STATUS_MSG,
+                &format!(
+                    "parameter checks specified: {} checks must be verified",
+                    { if self.param_checks_all { "all" } else { "some" } },
+                ),
+            );
 
-                                // if the result is still encapsulated in a Value, take it out
-                                while let zvariant::Value::Value(v) = field_value {
-                                    field_value = v;
-                                }
+            let severity;
+            let log_when;
+            let log_status;
+            let log_message;
 
-                                // now we should be ready for actual testing
-                                match &ck.value {
-                                    ParameterCheckValue::Boolean(b) => {
-                                        if ck.operator == ParamCheckOperator::Equal {
-                                            match field_value {
-                                                zvariant::Value::Bool(v) => {
-                                                    if *b == *v {
-                                                        verified = true;
-                                                        if !self.param_checks_all {
-                                                            break 'params;
-                                                        }
-                                                    } else {
-                                                        verified = false;
-                                                        if self.param_checks_all {
-                                                            break 'params;
-                                                        }
-                                                    }
-                                                }
-                                                e => {
-                                                    self.log(
-                                                        LogType::Warn,
-                                                        LOG_WHEN_PROC,
-                                                        LOG_STATUS_FAIL,
-                                                        &format!("mismatched result type: boolean expected (got `{e:?}`)"),
-                                                    );
-                                                    verified = false;
-                                                    break;
-                                                }
-                                            }
-                                        } else if ck.operator == ParamCheckOperator::Contains {
-                                            match field_value {
-                                                zvariant::Value::Array(_) => {
-                                                    if _contained_in(b, field_value) {
-                                                        verified = true;
-                                                        if !self.param_checks_all {
-                                                            break 'params;
-                                                        }
-                                                    } else {
-                                                        verified = false;
-                                                        if self.param_checks_all {
-                                                            break 'params;
-                                                        }
-                                                    }
-                                                }
-                                                _ => {
-                                                    verified = false;
-                                                    break;
-                                                }
-                                            }
-                                        } else if ck.operator == ParamCheckOperator::NotContains {
-                                            match field_value {
-                                                zvariant::Value::Array(_) => {
-                                                    if !_contained_in(b, field_value) {
-                                                        verified = true;
-                                                        if !self.param_checks_all {
-                                                            break 'params;
-                                                        }
-                                                    } else {
-                                                        verified = false;
-                                                        if self.param_checks_all {
-                                                            break 'params;
-                                                        }
-                                                    }
-                                                }
-                                                // incompatible checks should always yield false
-                                                _ => {
-                                                    verified = false;
-                                                    break;
-                                                }
-                                            }
-                                        } else {
-                                            self.log(
-                                                LogType::Warn,
-                                                LOG_WHEN_PROC,
-                                                LOG_STATUS_FAIL,
-                                                "invalid operator for boolean",
-                                            );
-                                            verified = false;
-                                            break;
-                                        }
-                                    }
-                                    ParameterCheckValue::Integer(i) => {
-                                        match field_value {
-                                            zvariant::Value::U8(v) => {
-                                                if _oper(&ck.operator, *i, *v as i64) {
-                                                    verified = true;
-                                                    if !self.param_checks_all {
-                                                        break 'params;
-                                                    }
-                                                } else {
-                                                    verified = false;
-                                                    if self.param_checks_all {
-                                                        break 'params;
-                                                    }
-                                                }
-                                            }
-                                            zvariant::Value::I16(v) => {
-                                                if _oper(&ck.operator, *i, *v as i64) {
-                                                    verified = true;
-                                                    if !self.param_checks_all {
-                                                        break 'params;
-                                                    }
-                                                } else {
-                                                    verified = false;
-                                                    if self.param_checks_all {
-                                                        break 'params;
-                                                    }
-                                                }
-                                            }
-                                            zvariant::Value::U16(v) => {
-                                                if _oper(&ck.operator, *i, *v as i64) {
-                                                    verified = true;
-                                                    if !self.param_checks_all {
-                                                        break 'params;
-                                                    }
-                                                } else {
-                                                    verified = false;
-                                                    if self.param_checks_all {
-                                                        break 'params;
-                                                    }
-                                                }
-                                            }
-                                            zvariant::Value::I32(v) => {
-                                                if _oper(&ck.operator, *i, *v as i64) {
-                                                    verified = true;
-                                                    if !self.param_checks_all {
-                                                        break 'params;
-                                                    }
-                                                } else {
-                                                    verified = false;
-                                                    if self.param_checks_all {
-                                                        break 'params;
-                                                    }
-                                                }
-                                            }
-                                            zvariant::Value::U32(v) => {
-                                                if _oper(&ck.operator, *i, *v as i64) {
-                                                    verified = true;
-                                                    if !self.param_checks_all {
-                                                        break 'params;
-                                                    }
-                                                } else {
-                                                    verified = false;
-                                                    if self.param_checks_all {
-                                                        break 'params;
-                                                    }
-                                                }
-                                            }
-                                            zvariant::Value::I64(v) => {
-                                                if _oper(&ck.operator, *i, *v as i64) {
-                                                    verified = true;
-                                                    if !self.param_checks_all {
-                                                        break 'params;
-                                                    }
-                                                } else {
-                                                    verified = false;
-                                                    if self.param_checks_all {
-                                                        break 'params;
-                                                    }
-                                                }
-                                            }
-                                            zvariant::Value::U64(v) => {
-                                                // lossy, however bigger numbers will just fail test
-                                                if _oper(&ck.operator, *i, *v as i64) {
-                                                    verified = true;
-                                                    if !self.param_checks_all {
-                                                        break 'params;
-                                                    }
-                                                } else {
-                                                    verified = false;
-                                                    if self.param_checks_all {
-                                                        break 'params;
-                                                    }
-                                                }
-                                            }
-                                            zvariant::Value::Array(_) => {
-                                                if ck.operator == ParamCheckOperator::Contains {
-                                                    if _contained_in(i, field_value) {
-                                                        verified = true;
-                                                        if !self.param_checks_all {
-                                                            break 'params;
-                                                        }
-                                                    } else {
-                                                        verified = false;
-                                                        if self.param_checks_all {
-                                                            break 'params;
-                                                        }
-                                                    }
-                                                } else if ck.operator == ParamCheckOperator::NotContains {
-                                                    if !_contained_in(i, field_value) {
-                                                        verified = true;
-                                                        if !self.param_checks_all {
-                                                            break 'params;
-                                                        }
-                                                    } else {
-                                                        verified = false;
-                                                        if self.param_checks_all {
-                                                            break 'params;
-                                                        }
-                                                    }
-                                                } else {
-                                                    verified = false;
-                                                    break 'params;
-                                                }
-                                            }
-                                            e => {
-                                                self.log(
-                                                    LogType::Warn,
-                                                    LOG_WHEN_PROC,
-                                                    LOG_STATUS_FAIL,
-                                                    &format!(
-                                                        "mismatched result type: {} expected (got `{e:?}`)",
-                                                        if ck.operator == ParamCheckOperator::Contains
-                                                            || ck.operator == ParamCheckOperator::NotContains { "container" }
-                                                        else { "integer" },
-                                                    ),
-                                                );
-                                                verified = false;
-                                                break 'params;
-                                            }
-                                        }
-                                    }
-                                    ParameterCheckValue::Float(f) => {
-                                        match field_value {
-                                            zvariant::Value::F64(v) => {
-                                                if _oper(&ck.operator, *f, *v) {
-                                                    verified = true;
-                                                    if !self.param_checks_all {
-                                                        break 'params;
-                                                    }
-                                                } else {
-                                                    verified = false;
-                                                    if self.param_checks_all {
-                                                        break 'params;
-                                                    }
-                                                }
-                                            }
-                                            zvariant::Value::Array(_) => {
-                                                if ck.operator == ParamCheckOperator::Contains {
-                                                    if _contained_in(f, field_value) {
-                                                        verified = true;
-                                                        if !self.param_checks_all {
-                                                            break 'params;
-                                                        }
-                                                    } else {
-                                                        verified = false;
-                                                        if self.param_checks_all {
-                                                            break 'params;
-                                                        }
-                                                    }
-                                                } else if ck.operator == ParamCheckOperator::NotContains {
-                                                    if !_contained_in(f, field_value) {
-                                                        verified = true;
-                                                        if !self.param_checks_all {
-                                                            break 'params;
-                                                        }
-                                                    } else {
-                                                        verified = false;
-                                                        if self.param_checks_all {
-                                                            break 'params;
-                                                        }
-                                                    }
-                                                } else {
-                                                    verified = false;
-                                                    break 'params;
-                                                }
-                                            }
-                                            e => {
-                                                self.log(
-                                                    LogType::Warn,
-                                                    LOG_WHEN_PROC,
-                                                    LOG_STATUS_FAIL,
-                                                    &format!(
-                                                        "mismatched result type: {} expected (got `{e:?}`)",
-                                                        if ck.operator == ParamCheckOperator::Contains
-                                                            || ck.operator == ParamCheckOperator::NotContains { "container" }
-                                                        else { "float" },
-                                                    ),
-                                                );
-                                                verified = false;
-                                                break 'params;
-                                            }
-                                        }
-                                    }
-                                    ParameterCheckValue::String(s) => {
-                                        if ck.operator == ParamCheckOperator::Equal {
-                                            match field_value {
-                                                zvariant::Value::Str(v) => {
-                                                    if *s == v.to_string() {
-                                                        verified = true;
-                                                        if !self.param_checks_all {
-                                                            break 'params;
-                                                        }
-                                                    } else {
-                                                        verified = false;
-                                                        if self.param_checks_all {
-                                                            break 'params;
-                                                        }
-                                                    }
-                                                }
-                                                zvariant::Value::ObjectPath(v) => {
-                                                    if *s == v.to_string() {
-                                                        verified = true;
-                                                        if !self.param_checks_all {
-                                                            break 'params;
-                                                        }
-                                                    } else {
-                                                        verified = false;
-                                                        if self.param_checks_all {
-                                                            break 'params;
-                                                        }
-                                                    }
-                                                }
-                                                e => {
-                                                    self.log(
-                                                        LogType::Warn,
-                                                        LOG_WHEN_PROC,
-                                                        LOG_STATUS_FAIL,
-                                                        &format!("mismatched result type: string expected (got `{e:?}`)"),
-                                                    );
-                                                    verified = false;
-                                                    break 'params;
-                                                }
-                                            }
-                                        } else if ck.operator == ParamCheckOperator::NotEqual {
-                                            match field_value {
-                                                zvariant::Value::Str(v) => {
-                                                    if *s != v.to_string() {
-                                                        verified = true;
-                                                        if !self.param_checks_all {
-                                                            break 'params;
-                                                        }
-                                                    } else {
-                                                        verified = false;
-                                                        if self.param_checks_all {
-                                                            break 'params;
-                                                        }
-                                                    }
-                                                }
-                                                zvariant::Value::ObjectPath(v) => {
-                                                    if *s != v.to_string() {
-                                                        verified = true;
-                                                        if !self.param_checks_all {
-                                                            break 'params;
-                                                        }
-                                                    } else {
-                                                        verified = false;
-                                                        if self.param_checks_all {
-                                                            break 'params;
-                                                        }
-                                                    }
-                                                }
-                                                e => {
-                                                    self.log(
-                                                        LogType::Warn,
-                                                        LOG_WHEN_PROC,
-                                                        LOG_STATUS_FAIL,
-                                                        &format!("mismatched result type: string expected (got `{e:?}`)"),
-                                                    );
-                                                    verified = false;
-                                                    break 'params;
-                                                }
-                                            }
-                                        } else if ck.operator == ParamCheckOperator::Contains {
-                                            if _contained_in(s, field_value) {
-                                                verified = true;
-                                                if !self.param_checks_all {
-                                                    break 'params;
-                                                }
-                                            } else {
-                                                verified = false;
-                                                if self.param_checks_all {
-                                                    break 'params;
-                                                }
-                                            }
-                                        } else if ck.operator == ParamCheckOperator::NotContains {
-                                            if !_contained_in(s, field_value) {
-                                                verified = true;
-                                                if !self.param_checks_all {
-                                                    break 'params;
-                                                }
-                                            } else {
-                                                verified = false;
-                                                if self.param_checks_all {
-                                                    break 'params;
-                                                }
-                                            }
-                                        } else {
-                                            self.log(
-                                                LogType::Warn,
-                                                LOG_WHEN_PROC,
-                                                LOG_STATUS_FAIL,
-                                                "invalid operator for string",
-                                            );
-                                            verified = false;
-                                            break 'params;
-                                        }
-                                    }
-                                    ParameterCheckValue::Regex(re) => {
-                                        if ck.operator == ParamCheckOperator::Match {
-                                            match field_value {
-                                                zvariant::Value::Str(v) => {
-                                                    if re.is_match(v.as_str()) {
-                                                        verified = true;
-                                                        if !self.param_checks_all {
-                                                            break 'params;
-                                                        }
-                                                    } else {
-                                                        verified = false;
-                                                        if self.param_checks_all {
-                                                            break 'params;
-                                                        }
-                                                    }
-                                                }
-                                                zvariant::Value::ObjectPath(v) => {
-                                                    if re.is_match(v.as_str()) {
-                                                        verified = true;
-                                                        if !self.param_checks_all {
-                                                            break 'params;
-                                                        }
-                                                    } else {
-                                                        verified = false;
-                                                        if self.param_checks_all {
-                                                            break 'params;
-                                                        }
-                                                    }
-                                                }
-                                                e => {
-                                                    self.log(
-                                                        LogType::Warn,
-                                                        LOG_WHEN_PROC,
-                                                        LOG_STATUS_FAIL,
-                                                        &format!("mismatched result type: string expected (got `{e:?}`)"),
-                                                    );
-                                                    verified = false;
-                                                    break 'params;
-                                                }
-                                            }
-                                        } else {
-                                            self.log(
-                                                LogType::Warn,
-                                                LOG_WHEN_PROC,
-                                                LOG_STATUS_FAIL,
-                                                "invalid operator for regular expression",
-                                            );
-                                            verified = false;
-                                            break 'params;
-                                        }
-                                    }
-                                }
-                            }
-                            _ => {
-                                self.log(
-                                    LogType::Warn,
-                                    LOG_WHEN_PROC,
-                                    LOG_STATUS_FAIL,
-                                    "could not retrieve field index: first index must be integer",
-                                );
-                                verified = false;
-                                break;
-                            }
-                        }
-
-                    } else {
-                        self.log(
-                            LogType::Warn,
-                            LOG_WHEN_PROC,
-                            LOG_STATUS_FAIL,
-                            "could not retrieve parameter: missing argument number",
-                        );
-                        verified = false;
-                        break;
-                    }
-                }
-            } else {
-                self.log(
-                    LogType::Warn,
-                    LOG_WHEN_PROC,
-                    LOG_STATUS_FAIL,
-                    "could not retrieve message body",
-                );
-            }
+            (
+                verified,
+                severity,
+                log_when,
+                log_status,
+                log_message,
+            ) = dbus_check_message(&message.unwrap(), checks, self.param_checks_all);
+            self.log(severity, log_when, log_status, &log_message);
 
         } else {
             panic!("attempt to verify condition without initializing tests")
